description_matcher/requirements.txt
numpy==1.18.5
nltk==3.5
absl_py==0.9.0
tensorflow==2.3.0
sentence_transformers==0.2.6.1
tensorflow_hub==0.8.0
requests==2.22.0
gensim==3.8.3
Keras==2.4.2
lxml==4.5.1
scipy==1.4.1
Django==2.2.12
absl==0.0
djangorestframework==3.12.2
laserembeddings==1.1.0
scikit_learn==0.23.2
